+++ src/app/activities/page.tsx
@@ -1,34 +1,14 @@
 'use client';
 
-import { useSearchParams } from 'next/navigation';
+import { Suspense } from 'react';
 
-import AllActivities from '@/features/activities/components/all-activities';
-import BannerCarousel from '@/features/activities/components/banner-carousel';
-import BestActivities from '@/features/activities/components/best-activities';
-import Search from '@/features/activities/components/search';
-import SearchResults from '@/features/activities/components/search-result';
+import ActivitiesPageContent from '@/app/activities/page-content';
 
 const ActivitiesPage = () => {
-  const searchParams = useSearchParams();
-  const keyword = searchParams.get('search')?.trim() || '';
-  const isSearching = keyword.trim().length > 0;
-
   return (
-    <main className="bg-background flex w-full flex-col gap-10">
-      <div className="mx-auto w-full max-w-[120rem]">
-        <BannerCarousel />
-        <Search />
-
-        {isSearching ? (
-          <SearchResults keyword={keyword} />
-        ) : (
-          <>
-            <BestActivities />
-            <AllActivities />
-          </>
-        )}
-      </div>
-    </main>
+    <Suspense fallback={<div>로딩 중...</div>}>
+      <ActivitiesPageContent />
+    </Suspense>
   );
 };
